Updated scrape_mars.py file
I Updated the scrape_mars.py file to allow time, 5 sec, between scrapes to reduce scrape errors. I was getting errors when a scrape was not finished before another began. So I update the code with time.sleep(5) function. It does take longer to retrieve the data when reloading of course, however I cross my fingers it reduces the errors overall. I did multiple tests, only once did I get an error and that was when the JPL image was still in the transition of image rotation, hence I changed the time from 3 sec to 5 sec between scrapes.